breaking more functions out into separate files.  breaking tests into separate test files.
