+++ tests/unit/runners/test_salt_describe_host.py
@@ -1,0 +1,81 @@
+# pylint: disable=line-too-long
+import logging
+from unittest.mock import call
+from unittest.mock import MagicMock
+from unittest.mock import mock_open
+from unittest.mock import patch
+
+import pytest
+import salt.config  # pylint: disable=import-error
+import salt.runners.salt as salt_runner  # pylint: disable=import-error
+import saltext.salt_describe.runners.salt_describe as salt_describe_runner
+import saltext.salt_describe.runners.salt_describe_host as salt_describe_host_runner
+import yaml
+
+log = logging.getLogger(__name__)
+
+
+@pytest.fixture
+def configure_loader_modules():
+    module_globals = {
+        "__salt__": {"salt.execute": salt_runner.execute},
+        "__opts__": salt.config.DEFAULT_MASTER_OPTS.copy(),
+    }
+    return {
+        salt_describe_runner: module_globals,
+        salt_describe_host_runner: module_globals,
+    }
+
+
+def test_host(tmp_path):
+    """
+    test describe.host
+    """
+    host_list = {
+        "poc-minion": {
+            "comment-0": ["# Host addresses"],
+            "comment-1": ["# comment"],
+            "127.0.0.1": {"aliases": ["localhost"]},
+            "127.0.1.1": {"aliases": ["megan-precision5550"]},
+            "::1": {"aliases": ["localhost", "ip6-localhost", "ip6-loopback"]},
+            "ff02::1": {"aliases": ["ip6-allnodes"]},
+            "ff02::2": {"aliases": ["ip6-allrouters"]},
+        }
+    }
+
+    expected_content = {
+        "host_file_content_0": {"host.present": [{"ip": "127.0.0.1"}, {"names": ["localhost"]}]},
+        "host_file_content_1": {
+            "host.present": [{"ip": "127.0.1.1"}, {"names": ["megan-precision5550"]}]
+        },
+        "host_file_content_2": {
+            "host.present": [
+                {"ip": "::1"},
+                {"names": ["localhost", "ip6-localhost", "ip6-loopback"]},
+            ]
+        },
+        "host_file_content_3": {"host.present": [{"ip": "ff02::1"}, {"names": ["ip6-allnodes"]}]},
+        "host_file_content_4": {"host.present": [{"ip": "ff02::2"}, {"names": ["ip6-allrouters"]}]},
+    }
+
+    expected_sls_write = yaml.dump(expected_content)
+    expected_include_write = yaml.dump({"include": ["poc-minion.host"]})
+    expected_calls = [
+        call().write(expected_sls_write),
+        call().write(expected_include_write),
+    ]
+
+    with patch.dict(
+        salt_describe_runner.__salt__, {"salt.execute": MagicMock(return_value=host_list)}
+    ):
+        with patch.dict(
+            salt_describe_runner.__salt__,
+            {"config.get": MagicMock(return_value=[tmp_path])},
+        ):
+            with patch("os.listdir", return_value=["host.sls"]):
+                with patch("salt.utils.files.fopen", mock_open()) as open_mock:
+                    assert salt_describe_host_runner.host("minion") is True
+                    open_mock.assert_has_calls(expected_calls, any_order=True)
+
+
+# pylint: enable=line-too-long
