+++ tests/unit/runners/test_salt_describe_pkg.py
@@ -1,0 +1,59 @@
+# pylint: disable=line-too-long
+import logging
+from unittest.mock import call
+from unittest.mock import MagicMock
+from unittest.mock import mock_open
+from unittest.mock import patch
+
+import pytest
+import salt.config  # pylint: disable=import-error
+import salt.runners.salt as salt_runner  # pylint: disable=import-error
+import saltext.salt_describe.runners.salt_describe as salt_describe_runner
+import saltext.salt_describe.runners.salt_describe_pkg as salt_describe_pkg_runner
+
+log = logging.getLogger(__name__)
+
+
+@pytest.fixture
+def configure_loader_modules():
+    module_globals = {
+        "__salt__": {"salt.execute": salt_runner.execute},
+        "__opts__": salt.config.DEFAULT_MASTER_OPTS.copy(),
+    }
+    return {
+        salt_describe_runner: module_globals,
+        salt_describe_pkg_runner: module_globals,
+    }
+
+
+def test_pkg():
+    pkg_list = {
+        "minion": {
+            "pkg1": "0.1.2-3",
+            "pkg2": "1.2rc5-3",
+            "pkg3": "2.3.4-5",
+            "pk4": "3.4-5",
+            "pkg5": "4.5.6-7",
+        }
+    }
+
+    expected_calls = [
+        call().write(
+            "installed_packages:\n  pkg.installed:\n  - pkgs:\n    - pkg1: 0.1.2-3\n    - pkg2: 1.2rc5-3\n    - pkg3: 2.3.4-5\n    - pk4: 3.4-5\n    - pkg5: 4.5.6-7\n"
+        ),
+        call().write("include:\n- minion.pkgs\n"),
+    ]
+    with patch.dict(
+        salt_describe_pkg_runner.__salt__, {"salt.execute": MagicMock(return_value=pkg_list)}
+    ):
+        with patch.dict(
+            salt_describe_pkg_runner.__salt__,
+            {"config.get": MagicMock(return_value=["/srv/salt", "/srv/spm/salt"])},
+        ):
+            with patch("os.listdir", return_value=["pkgs.sls"]):
+                with patch("salt.utils.files.fopen", mock_open()) as open_mock:
+                    assert salt_describe_pkg_runner.pkg("minion") is True
+                    open_mock.assert_has_calls(expected_calls, any_order=True)
+
+
+# pylint: enable=line-too-long
